% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{decode}
\alias{decode}
\title{Fonction qui gere l'affichage}
\usage{
decode(nb)
}
\arguments{
\item{nb}{un nombre}
}
\value{
liste de nombres et une de leurs decompositions
}
\description{
Fonction qui gere l'affichage
}
\examples{
decode(nombres_dec()[[1]])
}
