% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveplynR.R
\name{derive}
\alias{derive}
\title{derive}
\usage{
derive(p)
}
\arguments{
\item{p}{objet de la classe plynR}
}
\description{
cette fonction implémente la méthode derive
permettant de deriver les polynômes la classe plynR.
}
\examples{
p=plynR(coefs=c(2,5,3))
derive(p)
}
