% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft1_iter.R
\name{dft1_iter}
\alias{dft1_iter}
\title{dft1_iter implemente la version iterative de la transformee de Fourier discrete.}
\usage{
dft1_iter(k, x)
}
\arguments{
\item{k}{un nombre naturel}

\item{x}{un vecteur complexe}
}
\value{
Xk le nombre complexe correspondant
}
\description{
dft1_iter implemente la version iterative de la transformee de Fourier discrete.
}
\examples{
dft1_iter(2,c(3+1i,7+4i))
dft1_iter(5,c(5+2i,9+1i))

}
