% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft1_matrix.R
\name{dft1_matrix}
\alias{dft1_matrix}
\title{dft1_matrix implemente la version matricielle de la transformee de Fourier discrete en faisant usage de la fonction crossprod}
\usage{
dft1_matrix(k, x)
}
\arguments{
\item{k}{un nombre naturel}

\item{x}{un vecteur complexe}
}
\value{
Une matrice sur la base d'un vecteur complexe x et un nombre naturel k.
}
\description{
dft1_matrix implemente la version matricielle de la transformee de Fourier discrete en faisant usage de la fonction crossprod
}
\examples{
dft1_matrix(2,c(3+1i,7+4i))
dft1_matrix (5,c(5+2i,9+1i))

}
