% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft_factory.R
\name{dft_factory}
\alias{dft_factory}
\title{dft_factory}
\usage{
dft_factory(F, x)
}
\arguments{
\item{F}{une fonction.}

\item{x}{un vecteur complexe.}
}
\value{
un vecteur complexe
}
\description{
dft_factory implemente la transformee de Fourier discrete en faisant usage des versions soit,naive,iterative ou matricielle
}
\examples{
dft_factory(dft1_naive,c(3+1i,7+4i))
dft_factory(dft1_iter,c(3+1i,7+4i))
dft_factory(dft1_matrix,c(3+1i,7+4i))

}
