% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft_iter.R
\name{dft_iter}
\alias{dft_iter}
\title{dft_iter  implemente la transformee de Fourier discrete naive en faisant usage du "constructeur dft_factory
de la version iterative dft1_iter.}
\usage{
dft_iter(x)
}
\arguments{
\item{x}{un vecteur complexe}
}
\value{
vecteur complexe correspondanst à (Xk)
}
\description{
dft_iter  implemente la transformee de Fourier discrete naive en faisant usage du "constructeur dft_factory
de la version iterative dft1_iter.
}
\examples{
dft_iter (c(3+1i,7+4i))
dft_iter (c(5+2i,9+1i))

}
