% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft_matrix.R
\name{dft_matrix}
\alias{dft_matrix}
\title{dft_matrix}
\usage{
dft_matrix(x)
}
\arguments{
\item{x}{un vecteur complexe}
}
\value{
vecteur complexe correspondanst à (Xk)
}
\description{
dft_matrix   implemente la transformee de Fourier discrete naive en faisant usage du "constructeur dft_factory
de la version matricielle dft_matrix.
}
\examples{
dft_matrix (c(3+1i,7+4i))
dft_matrix (c(5+2i,9+1i))

}
