% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft_naive.R
\name{dft_naive}
\alias{dft_naive}
\title{dft_naive implemente la transformee de Fourier discrete naive en faisant usage du "constructeur dft_factory
de la version naive dft1_naive.}
\usage{
dft_naive(x)
}
\arguments{
\item{x}{un vecteur complexe}
}
\value{
vecteur complexe correspondanst à (Xk)
}
\description{
dft_naive implemente la transformee de Fourier discrete naive en faisant usage du "constructeur dft_factory
de la version naive dft1_naive.
}
\examples{
dft_naive(c(3+1i,7+4i))
dft_naive(c(5+2i,9+1i))

}
