% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_ct2.R
\name{fftct2}
\alias{fftct2}
\title{fftct2}
\usage{
fftct2(x)
}
\arguments{
\item{x}{un vecteur complexe}
}
\value{
le vecteur complexe correspondant à Xk.
}
\description{
la fonction fftct2 implemente la transformee de Fourier de Cooley Tukey de base  en utlisant la methode diviser pour conquerir
}
\examples{
fftct2(c(1+1i,4+7i,6+9i,20+3i))
fftct2(c(3+1i,7+4i))

}
