% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib_mulPos_rec.R
\name{fib_mulPos_rec}
\alias{fib_mulPos_rec}
\title{fib_mulPos_rec implemente la suite de Fibonnacci multiplicative positionnelle}
\usage{
fib_mulPos_rec(k)
}
\arguments{
\item{k}{un nombre entier positif.}
}
\value{
kieme terme de la suite de Fibonnacci multiplicative positionnelle.
}
\description{
fib_mulPos_rec implemente la suite de Fibonnacci multiplicative positionnelle
}
\examples{
fib_mulPos_rec(5)
fib_mulPos_rec(14)


}
