% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foisplynR.R
\name{fois}
\alias{fois}
\title{fois}
\usage{
fois(p, q)
}
\arguments{
\item{p}{objet de la classe plynR}

\item{q}{objet de la classe plynR}
}
\description{
la methode fois implémente la méthode fois(p, q)
de permettant de multiplier les polynômes la classe plynR.
}
\examples{
p=plynR(coefs=c(2,5,3))
q=plynR(coefs=c(1,4,7))
fois(p,q)
}
