% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moinsplynR.R
\name{moins}
\alias{moins}
\title{moins}
\usage{
moins(p, q)
}
\arguments{
\item{p}{objet de la classe plynR}

\item{q}{objet de la classe plynR}
}
\description{
la methode Implémente la méthode moins(p, q)
de permettant de sosutraire les polynômes la classe plynR
}
\examples{
p=plynR(coefs=c(2,5,3))
q=plynR(coefs=c(1,4,7))
moins(p,q)
}
