% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombres_dec.R
\name{nombres_dec}
\alias{nombres_dec}
\title{nombres_dec  est une fonction pour calculez le plus de termes possibles compris entre −1e10 et 1e10 tel que pour chacun des
termes calculés, il soit possible de retrouver sa «décomposition» facilement et rapidement.}
\usage{
nombres_dec(
  chiffres = c(2, 2, 5, 6, 6, 6, 7, 9),
  boundaries = c(-1e+10, 1e+10),
  nb_iters = 1e+05
)
}
\arguments{
\item{chiffres}{liste de nombres}

\item{boundaries}{intervalle de valeurs admissibles}

\item{nb_iters}{nombre d'iterations}
}
\value{
une liste des listes de nombres,d'operateurs et de candidats
}
\description{
nombres_dec  est une fonction pour calculez le plus de termes possibles compris entre −1e10 et 1e10 tel que pour chacun des
termes calculés, il soit possible de retrouver sa «décomposition» facilement et rapidement.
}
\examples{
nombres_dec()
}
