% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusplynR.R
\name{plus}
\alias{plus}
\title{plus}
\usage{
plus(p, q)
}
\arguments{
\item{p}{objet de la classe plynR}

\item{q}{objet de la classe plynR}
}
\description{
la methode Implémente la méthode plus(p,q)
de permettant d’additionner les polynômes la classe plynR
}
\examples{
p=plynR(coefs=c(2,5,3))
q=plynR(coefs=c(1,4,7))
plus(p,q)
}
