% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printplynR.R
\name{print}
\alias{print}
\title{print}
\usage{
print(x)
}
\arguments{
\item{x}{objet de la classe plynR}
}
\description{
la methode print(x) pour la classe plynR permettant d’afficher
le polynôme x a l’écran en faisant appel à la fonction cat
}
\examples{
x=plynR(coefs=c(2,5,3))
print(x)
}
