% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racinesplynR.R
\name{racines}
\alias{racines}
\title{Definition la methode Implémente la méthode racines.plynR
de permettant de trouver les racines les polynômes la classe plynR.}
\usage{
racines(p)
}
\arguments{
\item{p}{objet de la classe plynR}
}
\description{
Definition la methode Implémente la méthode racines.plynR
de permettant de trouver les racines les polynômes la classe plynR.
}
\examples{
p=plynR(coefs=c(2,5,3))
racines(p)
}
