% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_gen.R
\name{rand_gen}
\alias{rand_gen}
\title{rand_gen}
\usage{
rand_gen(x, trans = identity)
}
\arguments{
\item{x}{vecteur de nombres entiers de taille arbitraire}

\item{trans}{fonction de transformation}
}
\value{
sim fonction simulateur acceptant un seul argument n positif et tenant le compte de chacune des valeurs simulées
}
\description{
rand_gen retourne une fonction permettant de simuler des
}
\examples{
rand_gen(x=c(2,5,9,7,6),trans=identity)

}
