% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_stats.R
\name{rand_stats}
\alias{rand_stats}
\title{rand_stats}
\usage{
rand_stats(sim)
}
\arguments{
\item{sim}{est un simulateur}
}
\value{
un vecteur contenant la proportion de chacun des nombres générés jusqu’à maintenant
}
\description{
rand_stats est une fonction retournant un vecteur contenant la proportion de chacun des nombres générés jusqu’à maintenant
}
\examples{
sim=rand_gen(x=c(2,5,9,7,6),trans=identity)
sim(20)
rand_stats(sim)
}
