% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_sum.R
\name{rand_sum}
\alias{rand_sum}
\title{rand_sum}
\usage{
rand_sum(sim1, sim2)
}
\arguments{
\item{sim1}{un simulateur}

\item{sim2}{un simulateur}
}
\value{
un simulateur  simulant proportionnellement à x1 + x2.
}
\description{
une fonction permettant de faire la somme de deux simulateurs et
simulant des données proportionnellement aux vecteurs x1 et x2 respectivement et tel que Si x1 et x2
ne sont pas de la même longueur, on supposera que les valeurs absentes sont 0.
}
\examples{
sim1=rand_gen(x=c(2,5))
sim2=rand_gen(x=c(1,7,9))
rand_sum(sim1,sim2)
}
