% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_trans.R
\name{rand_trans}
\alias{rand_trans}
\title{rand_trans}
\usage{
rand_trans(sim, trans)
}
\arguments{
\item{sim}{le simulateur genere par rand_gen}

\item{trans}{la transformation est une fonction bijective à image dans les entiers positifs}
}
\value{
une fonction (simulateur) permettant de simuler de la distribution originale de sim transformée
}
\description{
rand_trans est une  fonction permettant d’appliquer une transformation à un simulateur produit
}
\examples{
rand_trans(rand_gen(x=c(2,5)),trans=identity)(3)
}
