% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryplynR.R
\name{summary}
\alias{summary}
\title{summary}
\usage{
summary(p)
}
\arguments{
\item{p}{objet de la classe plynR}
}
\description{
summary implémente la méthode summary.plynR.
permettant de trouver les racines les polynômes la classe plynR.
}
\examples{
p=plynR(coefs=c(2,5,3))
racines(p)
}
