% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{decode}
\alias{decode}
\title{res}
\usage{
decode(nb)
}
\arguments{
\item{nb}{un nombre}
}
\value{
liste de nombres et une de leurs decompositions
}
\description{
res est une onction qui gere l'affichage des nombres decomposes par nombres_dec
}
\examples{
decode(nombres_dec()[[1]])
}
