% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveplynR.R
\name{derive}
\alias{derive}
\title{Definition la methode Implémente la méthode derive(p)
de permettant de multiplier les polynômes la classe plynR.}
\usage{
derive(p)
}
\arguments{
\item{p}{objet de la classe plynR}
}
\description{
Definition la methode Implémente la méthode derive(p)
de permettant de multiplier les polynômes la classe plynR.
}
\examples{
p=plynR(coefs=c(2,5,3))
derive(p)
}
